samples/2_otel_cliapp_notool/pyproject.toml
[project]
name = "2_otel_cliapp_notool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "langchain>=1.0.7",
    "langchain-aws>=1.0.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
    "opentelemetry-exporter-otlp-proto-http>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "uvicorn>=0.38.0",
]
